+++ src/main/kotlin/ru/itmo/stand/cache/repository/TermRepository.kt
@@ -7,17 +7,18 @@ import org.springframework.stereotype.Repository
 class TermRepository(
     redisTemplate: RedisTemplate<String, Any>,
 ) {
-    private val hashOperations = redisTemplate.opsForHash<Double, String>()
+    private val hashOperations = redisTemplate.opsForHash<Float, String>()
     private val mapName = "term"
 
-    fun saveTerm(key: Double, value: String) {
+    fun saveTerm(key: Float, value: String) {
         hashOperations.put(mapName, key, value);
     }
-    fun saveTerms(terms: MutableMap<Double, String>) {
+
+    fun saveTerms(terms: MutableMap<Float, String>) {
         hashOperations.putAll(mapName, terms);
     }
     
-    fun getTerm(key: Double): String? {
+    fun getTerm(key: Float): String? {
         return hashOperations.get(mapName, key)
     }
 }
